model: Merge LicenseView tests into single class
The abstract class is not required anymore with the removal of the
`licenses` function and the related test in the previous commit.
